fixed footer issue
