+++ src/js/components/ContactFooter.jsx
@@ -1,27 +1,9 @@
 import React from 'react';
 import Firebase from 'firebase';
 
-var checkLocalStorage = function() {
-  var barName = localStorage.getItem('havaBarName');
-  if(barName !== null) {
-    return;
-  } else {
-    navigateToPage('/');
-  }
-}
-
-var navigateToPage = (pageUrl) => {
-  window.location = pageUrl;
-}
-
 var ContactFooter = React.createClass({
-  componentWillMount: function() {
-    checkLocalStorage();
-  },
-
   handleClickOnFooter: function(){
     var _this = this;
-    console.log('THIS!!!', _this);
     window.location.assign(this.props.navigateTo);
   },
